+++ src/pages/Detail.jsx
@@ -54,7 +54,9 @@ function Detail() {
   }, [commentList || post]);
 
   const deletePostHandler = (id) => {
-    dispatch(__deletePost(id));
+    if (window.confirm("삭제하시겠습니까?")) {
+      dispatch(__deletePost(id));
+    }
   };
 
   const donePostHandler = (item) => {
